requirements.txt
langgraph>=0.2.0
langchain-ollama>=0.2.0
pytest>=7.0.0
python-dotenv>=1.0.0
